Updated amplify config to auto fix gqlconfig that was created wrong
